refactor: api 타입스크립트 분리
